[Reviewer: Rob] Tie removing the data dir and removing from cluster together

+++ debian/clearwater-etcd.init.d
@@ -54,6 +54,7 @@
 DESC="etcd"
 NAME=clearwater-etcd
 DATA_DIR=/var/lib/$NAME
+JOINED_CLUSTER_SUCCESSFULLY=$DATA_DIR/clustered_successfully
 PIDFILE=/var/run/$NAME/$NAME.pid
 DAEMON=/usr/bin/etcd
 DAEMONWRAPPER=/usr/bin/etcdwrapper
@@ -153,15 +154,7 @@ join_cluster()
         # We need a temp file to deal with the environment variables.
         TEMP_FILE=$(mktemp)
 
-        # Build the client list based on $etcd_cluster, each entry is simply
-        # <IP>:<port> using the client port. Replace commas with whitespace,
-        # then split on whitespace (to cope with etcd_cluster values that have
-        # spaces)
-        export ETCDCTL_PEERS=
-        for server in ${etcd_cluster//,/ }
-        do
-            ETCDCTL_PEERS="$server:4000,$ETCDCTL_PEERS"
-        done
+        setup_etcdctl_peers
 
         # Check to make sure the cluster we want to join is healthy.
         # If it's not, don't even try joining (it won't work, and may
@@ -177,9 +170,12 @@ join_cluster()
         /usr/bin/etcdctl member add $ETCD_NAME http://$advertisement_ip:2380
         if [[ $? != 0 ]]
         then
+          local_member_id=$(/usr/bin/etcdctl member list | grep -F -w "http://$advertisement_ip:2380" | grep -o -E "^[^:]*" | grep -o "^[^[]\+")
+          /usr/bin/etcdctl member remove $local_member_id
           echo "Failed to add local node to cluster"
           exit 2
         fi
+
         ETCD_INITIAL_CLUSTER=$(/usr/share/clearwater/bin/get_etcd_initial_cluster.py $advertisement_ip $etcd_cluster)
 
         CLUSTER_ARGS="--initial-cluster $ETCD_INITIAL_CLUSTER
@@ -202,7 +198,8 @@ join_cluster()
 #
 join_or_create_cluster()
 {
-        if [[ $etcd_cluster =~ (^|,)$advertisement_ip(,|$) ]]
+        if [[ $etcd_cluster =~ (^|,)$advertisement_ip(,|$) ]] &&
+           [[ ! -f $JOINED_CLUSTER_SUCCESSFULLY ]]
         then
           create_cluster
         else
@@ -216,6 +213,7 @@ wait_for_etcd()
         start_time=$(date +%s)
         while true; do
           if nc -z $listen_ip 4000; then
+            touch $JOINED_CLUSTER_SUCCESSFULLY
             break;
           else
             current_time=$(date +%s)
@@ -240,12 +238,15 @@ verify_etcd_health()
         # The [unstarted] is only present while the member hasn't fully joined the etcd cluster
         setup_etcdctl_peers
         member_list=$(/usr/bin/etcdctl member list)
-        local_member_id=$(echo $member_list | grep -F -w "http://$local_ip:2380" | grep -o -E "^[^:]*" | grep -o "^[^[]\+")
-        unstarted_member_id=$(echo $member_list | grep -F -w "http://$local_ip:2380" | grep "unstarted")
+        local_member_id=$(echo "$member_list" | grep -F -w "http://$local_ip:2380" | grep -o -E "^[^:]*" | grep -o "^[^[]\+")
+        unstarted_member_id=$(echo "$member_list" | grep -F -w "http://$local_ip:2380" | grep "unstarted")
         if [[ $unstarted_member_id != '' ]]
         then
           /usr/bin/etcdctl member remove $local_member_id
-          rm -rf $DATA_DIR/$advertisement_ip
+          if [[ $? == 0 ]]
+          then
+            rm -rf $DATA_DIR/$advertisement_ip
+          fi
         fi
 
         if [[ -e $DATA_DIR/$advertisement_ip ]]
@@ -260,7 +261,10 @@ verify_etcd_health()
           if [[ $rc != 0 ]]
           then
             /usr/bin/etcdctl member remove $local_member_id
-            rm -rf $DATA_DIR/$advertisement_ip
+            if [[ $? == 0 ]]
+            then
+              rm -rf $DATA_DIR/$advertisement_ip
+            fi
           fi
         fi
 }
@@ -331,6 +335,8 @@ do_start()
 
 do_rebuild()
 {
+        rm -f $JOINED_CLUSTER_SUCCESSFULLY
+
         # Return
         #   0 if daemon has been started
         #   1 if daemon was already running
